+++ src/modules/core/provider/FunctionTreeItem.ts
@@ -1,7 +1,7 @@
 import { TreeItem, TreeItemCollapsibleState } from "vscode"
 import { PoolConnectionConfig } from "../types"
 
-export class ProcedureTreeItem extends TreeItem {
+export class FunctionTreeItem extends TreeItem {
     public connectionConfig: PoolConnectionConfig
 
     constructor(config: {
@@ -11,10 +11,10 @@ export class ProcedureTreeItem extends TreeItem {
         super(config.label, TreeItemCollapsibleState.None)
         // this.iconPath = new ThemeIcon('play')
         this.connectionConfig = config.connectionConfig
-        this.contextValue = 'procedure'
+        this.contextValue = 'function'
 
         this.command = {
-            command: 'SmileDB.openProcedure',
+            command: 'SmileDB.openFunction',
             title: 'Open Code',
             arguments: [config.connectionConfig, config.label]
         }