+++ src/modules/core/provider/TableTreeItem.ts
@@ -23,6 +23,4 @@ export class TableTreeItem extends TreeItem {
             arguments: [this.connectionConfig, this.table]
         }
     }
-
-
 }