+++ src/modules/mongodb/getDatabaseTreeChildren.ts
@@ -6,13 +6,13 @@ export async function getDatabaseTreeChildren(item: TreeItem): Promise<TreeItem[
     if (item instanceof PoolConnectionTreeItem) {
         // root
         const connection = getPoolConnection(item.connectionConfig) as MongoDBPoolConnection
-        const { rows, fields } = await connection.fetchDatabases()
-        return rows.map(row => new DatabaseTreeItem(
+        const databases = await connection.fetchDatabases()
+        return databases.map(database => new DatabaseTreeItem(
             {
                 ...item.connectionConfig,
                 connection: {
                     ...item.connectionConfig.connection,
-                    database: row[fields[0].name]?.toString() || '',
+                    database,
                 }
             },
         ))
@@ -20,20 +20,20 @@ export async function getDatabaseTreeChildren(item: TreeItem): Promise<TreeItem[
 
     else if (item instanceof DatabaseTreeItem) {
         const connection = getPoolConnection(item.connectionConfig) as MongoDBPoolConnection
-        const { rows } = await connection.fetchDatabaseStats()
-        const totalCollections = parseInt(rows[0].collections?.toString() || "0")
-        const totalViews = parseInt(rows[0].views?.toString() || "0")
+        const result = await connection.fetchDatabaseStats()
+        const totalCollections = result.collections
+        const totalViews = result.views
         return [
             new FolderTreeItem({
                 label: "collections",
-                contextValue: "collectionFolder",
+                contextValue: "collections",
                 connectionConfig: item.connectionConfig,
                 description: totalCollections.toString(),
                 state: totalCollections > 0 ? TreeItemCollapsibleState.Collapsed : TreeItemCollapsibleState.None
             }),
             new FolderTreeItem({
                 label: "views",
-                contextValue: "viewFolder",
+                contextValue: "views",
                 connectionConfig: item.connectionConfig,
                 description: totalViews.toString(),
                 state: totalViews > 0 ? TreeItemCollapsibleState.Collapsed : TreeItemCollapsibleState.None
@@ -41,11 +41,11 @@ export async function getDatabaseTreeChildren(item: TreeItem): Promise<TreeItem[
         ]
     }
 
-    else if (item instanceof FolderTreeItem && item.contextValue === 'collectionFolder') {
+    else if (item instanceof FolderTreeItem && item.contextValue === 'collections') {
         return await loadAndCreateTreeItem(item, 'fetchTables', TableTreeItem)
     }
 
-    else if (item instanceof FolderTreeItem && item.contextValue === 'viewFolder') {
+    else if (item instanceof FolderTreeItem && item.contextValue === 'views') {
         return await loadAndCreateTreeItem(item, 'fetchViews', TableTreeItem)
     }